Intelligent Assistant/Services/DB_Dummy/requirements.txt
influxdb_client==1.36.0
pandas==1.3.5
PyYaml==6.0
